requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0,<2.0
numpy>=2.1,<3.0
transformers>=4.44,<5.0
torch>=2.3,<3.0
scipy>=1.13,<2.0
